Fix visible observed attribute overwriting DOM
This fixes an issue where using the 'visible' visibility option in a binding
will cause the value of the observed attribute to be set via the $.text()
jQuery method on the selected element.  An example of the behavior can be seen
here: http://jsbin.com/udapeb/8/edit

This patch sets updateView to false for 'visible' bindings.
Additionally, there is an existence check in the default text binding handler.

Tests updated to verify the content of the bound element.